feat: add pagination to product list

+++ src/main/java/com/example/regionaldelicacy/services/ProductService.java
@@ -1,20 +1,16 @@
 package com.example.regionaldelicacy.services;
 
-import java.util.List;
 import java.util.Optional;
-import java.util.stream.Collectors;
 
-import org.springframework.data.domain.PageRequest;
+import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
-import org.springframework.security.core.Authentication;
-import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
 
-import com.example.regionaldelicacy.dtos.CreateProductDto;
 import com.example.regionaldelicacy.dtos.FavoriteInfoDto;
+import com.example.regionaldelicacy.dtos.ProductCreateUpdateDto;
 import com.example.regionaldelicacy.exceptions.CategoryNotFoundException;
 import com.example.regionaldelicacy.exceptions.ProductNotFoundException;
+import com.example.regionaldelicacy.models.Category;
 import com.example.regionaldelicacy.models.Favorite;
 import com.example.regionaldelicacy.models.Product;
 import com.example.regionaldelicacy.models.User;
@@ -23,45 +19,48 @@ import com.example.regionaldelicacy.repositories.FavoriteRepository;
 import com.example.regionaldelicacy.repositories.ProductRepository;
 
 import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
 
 @Service
 @RequiredArgsConstructor
+@Slf4j
 public class ProductService {
+
     private final ProductRepository productRepository;
     private final CategoryRepository categoryRepository;
     private final FavoriteRepository favoriteRepository;
 
-    public List<Product> getAllProducts() {
-        return productRepository.findAll();
+    public Page<Product> getAllProducts(Pageable pageable) {
+        return productRepository.findAll(pageable);
     }
 
-    public List<Product> getProductsByCategory(String categoryParam) {
+    public Page<Product> getProductsByCategory(String categoryParam, Pageable pageable) {
         try {
             Long categoryId = Long.parseLong(categoryParam);
-            return productRepository.findByCategoryCategoryId(categoryId);
+            return productRepository.findByCategoryId(categoryId, pageable);
         } catch (NumberFormatException e) {
-            return productRepository.findByCategoryNameIgnoreCase(categoryParam);
+            return productRepository.findByCategoryNameIgnoreCase(categoryParam, pageable);
         }
     }
 
-    public List<Product> searchProducts(String searchTerm, String categoryParam) {
+    public Page<Product> searchProducts(String searchTerm, String categoryParam, Pageable pageable) {
         if (searchTerm == null || searchTerm.trim().isEmpty()) {
-            return categoryParam == null ? getAllProducts() : getProductsByCategory(categoryParam);
+            return categoryParam == null ? getAllProducts(pageable) : getProductsByCategory(categoryParam, pageable);
         }
 
         String trimmedSearchTerm = searchTerm.trim();
 
         if (categoryParam == null) {
-            return productRepository.findByNameContainingIgnoreCaseAndAccent(trimmedSearchTerm);
+            return productRepository.findByNameContainingIgnoreCaseAndAccent(trimmedSearchTerm, pageable);
         }
 
         try {
             Long categoryId = Long.parseLong(categoryParam);
-            return productRepository.findByNameContainingIgnoreCaseAndAccentAndCategoryCategoryId(trimmedSearchTerm,
-                    categoryId);
+            return productRepository.findByNameContainingIgnoreCaseAndAccentAndCategoryId(trimmedSearchTerm,
+                    categoryId, pageable);
         } catch (NumberFormatException e) {
             return productRepository.findByNameContainingIgnoreCaseAndAccentAndCategoryNameIgnoreCase(trimmedSearchTerm,
-                    categoryParam);
+                    categoryParam, pageable);
         }
     }
 
@@ -70,45 +69,54 @@ public class ProductService {
         return product.orElseThrow(ProductNotFoundException::new);
     }
 
-    @Transactional
-    public Product saveProduct(CreateProductDto productDto) {
-        Product product = productDto.toProduct();
-        product.setCategory(categoryRepository.findById(productDto.categoryId())
+    public Product createProduct(ProductCreateUpdateDto productCreateDto) {
+        Product product = productCreateDto.toProduct();
+        product.setCategory(categoryRepository.findById(productCreateDto.getCategoryId())
                 .orElseThrow(CategoryNotFoundException::new));
         return productRepository.save(product);
     }
 
-    public void deleteProduct(Long id) {
+    public Product updateProductById(Long id, ProductCreateUpdateDto productUpdateDto) {
+        Product product = productRepository.findById(id).orElseThrow(ProductNotFoundException::new);
+        Category category = categoryRepository.findById(productUpdateDto.getCategoryId())
+                .orElseThrow(CategoryNotFoundException::new);
+
+        product.setName(productUpdateDto.getName());
+        product.setDescription(productUpdateDto.getDescription());
+        product.setPrice(productUpdateDto.getPrice());
+        product.setImageUrl(productUpdateDto.getImageUrl());
+        product.setBrand(productUpdateDto.getBrand());
+        product.setStock(productUpdateDto.getStock());
+        product.setCategory(category);
+
+        return productRepository.save(product);
+    }
+
+    public void deleteProductById(Long id) {
         productRepository.deleteById(id);
     }
 
-    public List<FavoriteInfoDto> getFavoriteProducts(Integer page, Integer size) {
-        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
-        User user = (User) authentication.getPrincipal();
-        Pageable paging = PageRequest.of(page, size);
-        List<Favorite> favoriteProducts = favoriteRepository.findByUserUserId(user.getUserId(), paging);
-        List<FavoriteInfoDto> favoriteDtos = favoriteProducts.stream().map(favoriteProduct -> {
-            return FavoriteInfoDto.fromFavorite(favoriteProduct);
-        }).collect(Collectors.toList());
+    public Page<FavoriteInfoDto> getFavoriteProducts(User user, Pageable pageable) {
+        Page<Favorite> favoriteProducts = favoriteRepository.findByUserUserId(user.getUserId(), pageable);
+        Page<FavoriteInfoDto> favoriteDtos = favoriteProducts.map(FavoriteInfoDto::fromFavorite);
         return favoriteDtos;
     }
 
-    @Transactional
-    public FavoriteInfoDto addFavoriteProduct(Long productId) {
-        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
-        User user = (User) authentication.getPrincipal();
+    public FavoriteInfoDto addFavoriteProduct(User user, Long productId) {
         Favorite favorite = favoriteRepository.findByUserUserIdAndProductProductId(user.getUserId(), productId);
         if (favorite == null) {
             favorite = Favorite.builder()
                     .user(user)
                     .product(getProductById(productId))
                     .build();
-            favoriteRepository.save(favorite);
+            favorite = favoriteRepository.save(favorite);
         }
         return FavoriteInfoDto.fromFavorite(favorite);
     }
 
-    public void removeFavoriteProduct(Long favoriteId) {
-        favoriteRepository.deleteById(favoriteId);
+    public void removeFavoriteProduct(User user, Long favoriteId) {
+        if (favoriteRepository.findByIdAndUserUserId(favoriteId, user.getUserId()).isPresent()) {
+            favoriteRepository.deleteById(favoriteId);
+        }
     }
 }
